Refactor form controls
