+++ src/components/settings-form/SettingsForm.tsx
@@ -1,96 +1,26 @@
-import React, { useState, MouseEvent } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 
 import Button from "@mui/material/Button";
-import InputLabel from "@mui/material/InputLabel";
-import FormControl from "@mui/material/FormControl";
-import { SelectChangeEvent } from "@mui/material/Select";
-import MenuItem from "@mui/material/MenuItem";
-import Select from "@mui/material/Select";
+
+import {
+  ButtonFormControl,
+  ActionButtonData,
+} from "./form-controls/ButtonFormControl";
+import {
+  SelectFormControl,
+  SelectFormControlData,
+} from "./form-controls/SelectFormControl";
 
 import { I18nCopy } from "../../types/app";
 
 import styles from "./SettingsForm.module.scss";
 
-interface EnumMap {
-  [key: string]: string;
-}
-
-type Options = { [key: string]: string };
-
-interface SettingsFormProps {
+export interface SettingsFormProps {
   copy: I18nCopy;
   selectFormControls: SelectFormControlData[];
   actionButtons: ActionButtonData[];
 }
-
-const renderSelectOptions = (
-  key: string,
-  enumOptions: EnumMap,
-  options: Options,
-) =>
-  Object.keys(enumOptions).map((value: string) => (
-    <MenuItem key={`${key}-options-item-${value}`} value={value}>
-      {options[value.toString().toLowerCase()]}
-    </MenuItem>
-  ));
-
-export interface SelectFormControlData {
-  key: string;
-  copy: I18nCopy;
-  value: string;
-  enum: EnumMap;
-  onChange: (e: SelectChangeEvent<string>) => void;
-  options: Options;
-}
-
-export interface ActionButtonData {
-  key: string;
-  copy: I18nCopy;
-  onClick: (e: MouseEvent) => Promise<void>;
-}
-
-const SelectFormControl = (item: SelectFormControlData) => (
-  <div
-    key={`form-control-item-${item.key}`}
-    className={styles.SettingsForm__formControl}
-  >
-    <FormControl>
-      <InputLabel id={`${item.key}-label`}>{item.copy.label}</InputLabel>
-      <Select
-        className={styles.Select}
-        labelId={`${item.key}-label`}
-        id={`${item.key}-select`}
-        value={item.value}
-        label={item.copy.label}
-        onChange={item.onChange}
-      >
-        {renderSelectOptions(item.key, item.enum, item.options)}
-      </Select>
-    </FormControl>
-  </div>
-);
-
-const ActionButton = (item: ActionButtonData) => {
-  const [loading, setLoading] = useState(false);
-
-  const onClick = async (e: MouseEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      await item.onClick(e);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <Button onClick={onClick} type="submit" color="primary" variant="contained">
-      {item.copy.label} {loading ? "⏳" : ""}
-    </Button>
-  );
-};
 
 const SettingsForm = (props: SettingsFormProps) => (
   <form className={styles["SettingsForm"]} noValidate autoComplete="off">
@@ -101,7 +31,7 @@ const SettingsForm = (props: SettingsFormProps) => (
     </div>
     <div className={styles["SettingsForm__actions"]}>
       {props.actionButtons.map((actionButton) => (
-        <ActionButton {...actionButton} />
+        <ButtonFormControl {...actionButton} />
       ))}
     </div>
     <div className={styles["SettingsForm__controls"]}>
